% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceIntervalCalculations.R
\name{confidence_ints}
\alias{confidence_ints}
\title{Confidence Interval Calculator}
\usage{
confidence_ints(data)
}
\arguments{
\item{data}{an array of numbers}
}
\description{
This function takes in an array of numbers and gives back mean, standard erro
r, t-score, margin error and the onfidence intervals.
}
