% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxplotDistributions.R
\name{distributions}
\alias{distributions}
\title{Plot distributions}
\usage{
distributions(dataset)
}
\arguments{
\item{dataset}{an array of numbers}
}
\description{
This function takes in an array of numbers and and plots a boxplot.
}
