% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrequencyCalculations.R
\name{freq}
\alias{freq}
\title{Find Frequencies and Plots Frequency Histograms}
\usage{
freq(dataset)
}
\arguments{
\item{dataset}{an array of numbers}
}
\value{
frequency of each value
}
\description{
This function takes in an array of numbers and gives the frequency of each va
lue as well as plots the frequency histogram
}
